% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_helpers.R
\name{tat_summary}
\alias{tat_summary}
\title{Get Turnaround Time (TAT) Summary}
\usage{
tat_summary(con, branch = "All", sector = "All")
}
\arguments{
\item{con}{A DBI connection object}

\item{branch}{Branch name to filter by (or "All")}

\item{sector}{Sector name to filter by (or "All")}
}
\value{
A dataframe with average TAT in hours by stage and branch
}
\description{
Get Turnaround Time (TAT) Summary
}
