% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_helpers.R
\name{vintage_curve}
\alias{vintage_curve}
\title{Calculate Vintage Delinquency Curves}
\usage{
vintage_curve(con)
}
\arguments{
\item{con}{A DBI connection object}
}
\value{
A dataframe with cohort_month, mob, and delinq_rate
}
\description{
Calculate Vintage Delinquency Curves
}
